alert-service/requirements.txt
kafka-python==2.0.2
requests==2.28.1
flask==2.0.1
werkzeug==2.0.3 